server/requirements.txt
kafka-python
fastapi
uvicorn
pydantic